xtheadmempair: Clarify atomicity for load/store-pair instructions
Load/store-pair instructions raise the question about atomicity
and exception behaviour. The current description does not define
the exact behaviour (and thus makes this an implementation detail).

This patch clarifies, that the instructions don't provide any
atomicity guarantees and that that exceptions may re-execute
the instruction causing duplicated memory transactions.

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>