Add intrinsics for xtheadvector extension
We are now adding XTheadVector C intrinsics that provide users
interfaces in the C language level to directly leverage XTheadVector.

XTheadVector has a huge overlap with RVV. So it is reasonable,
to reuse the RVV intrinsic API for XTheadVector as well
(when applicable) and extend the API by additional functions.