Introduce th.mxstatus CSR
To implement proper extension discovery, we need to specify a mechanism
to identify the availability of XThead* extension in a system.
This mechanism already exists with the th.mxstatus register's
THEADISAEE bit. Let's document this CSR.

Other bits than THEADISAEE are either 0 or defined as WPRI.
This allows to define them in other extensions (e.g. UCME is
defined as part of XTheadCmo and MAEE is defined as part of XTheadMaee).

The th.mxstatus CSR is defined to be available on all system's with a
mvendorid or 0x5B7.

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>