XTheadMae: Rename and cleanup memory attribute extension
We named the extension "MAEE" following the name of the enable bit
in the SxStatus CSR. However, the last 'E' stands for enable and should
therefore not be part of the name. Renaming an extension that has been
published is usually a no-go, but since we don't have any support
patches for "XTheadMaee" merged yet, we can get away with it.
Therefore, let's rename XTheadMaee to XTheadMae now.

Further cleanups:
* Streamline name to `XTheadMae`
* Simplify availability section (following other extensions)

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>